Address review nitpicks
- String byte counts accumulate in long, so no possible string length
  can overflow them.
- A new test pins per-flavor ceiling counting: 43 emoji are 258
  CESU-8 bytes and fail ClassiCube validation, though standard UTF-8
  would count only 172 and let them through.
- NbtString and NbtTagType docs name the per-flavor string encodings.
- NbtFile save and NbtWriter array, list, and tag docs mention
  rejection by enabled validation.
- Reworded a stale comment crediting NbtCodec alone with setting
  reader limits.

+++ fNbt/NbtBinaryWriter.cs
@@ -265,7 +265,7 @@ namespace fNbt {
 
 
         void WriteModifiedUtf8(string value) {
-            int numBytes = NbtStringCodec.GetModifiedByteCount(value);
+            long numBytes = NbtStringCodec.GetModifiedByteCount(value);
             if (useVarInt) {
                 WriteUnsignedVarInt32((uint)numBytes);
             } else {
